Rename pointer tracker file
