+++ utils/PointerTracker.ts
